backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0
pdfplumber==0.10.3
PyPDF2==3.0.1
google-genai==0.2.0
rank-bm25==0.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.23
sqlmodel==0.0.14
watchfiles==0.21.0
